add assert_match with deprecation warning
